poc-mp/src/test/java/com/xyz/testclient/model/TopicDwFeedEventPayloadTest.java
package com.xyz.testclient.model;

import com.xyz.ims.domain.model.TopicDwFeedEventPayload;
import com.xyz.testclient.TopicDwFeedEventPayloadFactory;
import org.junit.jupiter.api.Test;
import org.springframework.util.Assert;

import java.text.ParseException;

class TopicDwFeedEventPayloadTest {

    @Test
    void toJson() throws ParseException {
        TopicDwFeedEventPayload topicDwFeedEventPayload = TopicDwFeedEventPayloadFactory.generateOneRandomEmailOpenFeedEventPayload();
        String result = topicDwFeedEventPayload.toJson();
        Assert.isTrue(result != null && result.length() > 0);

        Assert.isTrue(result.indexOf("category") > 0);
        Assert.isTrue(result.indexOf("discussion") > 0);
        Assert.isTrue(result.indexOf("user") > 0);
        Assert.isTrue(result.indexOf("date") > 0);
        Assert.isTrue(result.indexOf("email") > 0);

    }
}